Fix CLI not to read space state when it's not needed
